+++ src/Logger/LogInterface.php
@@ -6,6 +6,7 @@ namespace ExtendsSoftware\ExaPHP\Logger;
 
 use DateTime;
 use ExtendsSoftware\ExaPHP\Logger\Priority\PriorityInterface;
+use Throwable;
 
 interface LogInterface
 {
@@ -38,6 +39,13 @@ interface LogInterface
     public function getMetaData(): array;
 
     /**
+     * Get throwable.
+     *
+     * @return Throwable|null
+     */
+    public function getThrowable(): ?Throwable;
+
+    /**
      * Return new log with message.
      *
      * @param string $message
@@ -64,4 +72,13 @@ interface LogInterface
      * @return LogInterface
      */
     public function andMetaData(string $key, mixed $value): LogInterface;
+
+    /**
+     * Return new log with throwable.
+     *
+     * @param Throwable|null $throwable
+     *
+     * @return LogInterface
+     */
+    public function withThrowable(Throwable $throwable = null): LogInterface;
 }