+++ src/Logger/Logger.php
@@ -8,6 +8,7 @@ use ExtendsSoftware\ExaPHP\Logger\Decorator\DecoratorInterface;
 use ExtendsSoftware\ExaPHP\Logger\Priority\PriorityInterface;
 use ExtendsSoftware\ExaPHP\Logger\Writer\WriterException;
 use ExtendsSoftware\ExaPHP\Logger\Writer\WriterInterface;
+use Throwable;
 
 class Logger implements LoggerInterface
 {
@@ -29,9 +30,13 @@ class Logger implements LoggerInterface
      * @inheritDoc
      * @throws WriterException When writer failed to write.
      */
-    public function log(string $message, PriorityInterface $priority = null, array $metaData = null): LoggerInterface
-    {
-        $log = new Log($message, $priority ?? null, null, $metaData ?? null);
+    public function log(
+        string $message,
+        PriorityInterface $priority = null,
+        Throwable $throwable = null,
+        array $metaData = null,
+    ): LoggerInterface {
+        $log = new Log($message, $priority, null, $throwable, $metaData);
         $log = $this->decorate($log);
 
         foreach ($this->writers as $writer) {