+++ src/Logger/LoggerInterface.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace ExtendsSoftware\ExaPHP\Logger;
 
 use ExtendsSoftware\ExaPHP\Logger\Priority\PriorityInterface;
+use Throwable;
 
 interface LoggerInterface
 {
@@ -15,9 +16,15 @@ interface LoggerInterface
      *
      * @param string                 $message
      * @param PriorityInterface|null $priority
+     * @param Throwable|null         $throwable
      * @param mixed[]|null           $metaData
      *
      * @return LoggerInterface
      */
-    public function log(string $message, PriorityInterface $priority = null, array $metaData = null): LoggerInterface;
+    public function log(
+        string $message,
+        PriorityInterface $priority = null,
+        Throwable $throwable = null,
+        array $metaData = null,
+    ): LoggerInterface;
 }