+++ tests/Logger/LogTest.php
@@ -1,9 +1,11 @@
 <?php
+
 declare(strict_types=1);
 
 namespace ExtendsSoftware\ExaPHP\Logger;
 
 use PHPUnit\Framework\TestCase;
+use Throwable;
 
 class LogTest extends TestCase
 {
@@ -17,6 +19,7 @@ class LogTest extends TestCase
      * @covers \ExtendsSoftware\ExaPHP\Logger\Log::getPriority()
      * @covers \ExtendsSoftware\ExaPHP\Logger\Log::getDateTime()
      * @covers \ExtendsSoftware\ExaPHP\Logger\Log::getMetaData()
+     * @covers \ExtendsSoftware\ExaPHP\Logger\Log::getThrowable()
      */
     public function testGetMethods(): void
     {
@@ -25,6 +28,7 @@ class LogTest extends TestCase
         $this->assertSame('Error!', $log->getMessage());
         $this->assertIsObject($log->getPriority());
         $this->assertIsObject($log->getDateTime());
+        $this->assertNull($log->getThrowable());
         $this->assertSame([], $log->getMetaData());
     }
 
@@ -36,23 +40,30 @@ class LogTest extends TestCase
      * @covers \ExtendsSoftware\ExaPHP\Logger\Log::__construct()
      * @covers \ExtendsSoftware\ExaPHP\Logger\Log::withMessage()
      * @covers \ExtendsSoftware\ExaPHP\Logger\Log::withMetaData()
+     * @covers \ExtendsSoftware\ExaPHP\Logger\Log::withThrowable()
      * @covers \ExtendsSoftware\ExaPHP\Logger\Log::getMessage()
      * @covers \ExtendsSoftware\ExaPHP\Logger\Log::getMetaData()
+     * @covers \ExtendsSoftware\ExaPHP\Logger\Log::getThrowable()
      */
     public function testWithMethods(): void
     {
+        $throwable = $this->createMock(Throwable::class);
+
         $log = new Log('Error!', metaData: [
             'foo' => 'bar',
         ]);
+
         $instance = $log
             ->withMessage('Exception!')
             ->withMetaData([
                 'bar' => 'baz',
-            ]);
+            ])
+            ->withThrowable($throwable);
 
         $this->assertNotSame($log, $instance);
         $this->assertIsObject($instance);
         $this->assertSame('Exception!', $instance->getMessage());
+        $this->assertSame($throwable, $instance->getThrowable());
         $this->assertSame(['bar' => 'baz'], $instance->getMetaData());
     }