+++ tests/Logger/LoggerTest.php
@@ -1,4 +1,5 @@
 <?php
+
 declare(strict_types=1);
 
 namespace ExtendsSoftware\ExaPHP\Logger;
@@ -7,6 +8,7 @@ use ExtendsSoftware\ExaPHP\Logger\Decorator\DecoratorInterface;
 use ExtendsSoftware\ExaPHP\Logger\Priority\PriorityInterface;
 use ExtendsSoftware\ExaPHP\Logger\Writer\WriterInterface;
 use PHPUnit\Framework\TestCase;
+use Throwable;
 
 class LoggerTest extends TestCase
 {
@@ -26,6 +28,7 @@ class LoggerTest extends TestCase
     public function testLog(): void
     {
         $priority = $this->createMock(PriorityInterface::class);
+        $throwable = $this->createMock(Throwable::class);
 
         $decorator = $this->createMock(DecoratorInterface::class);
         $decorator
@@ -39,23 +42,27 @@ class LoggerTest extends TestCase
         $writer
             ->expects($this->once())
             ->method('write')
-            ->with($this->callback(function (LogInterface $log) use ($priority) {
-                $this->assertSame('Error!', $log->getMessage());
-                $this->assertSame($priority, $log->getPriority());
-                $this->assertSame(['foo' => 'bar', 'baz' => 'qux'], $log->getMetaData());
+            ->with(
+                $this->callback(function (LogInterface $log) use ($priority, $throwable) {
+                    $this->assertSame('Error!', $log->getMessage());
+                    $this->assertSame($priority, $log->getPriority());
+                    $this->assertSame($throwable, $log->getThrowable());
+                    $this->assertSame(['foo' => 'bar', 'baz' => 'qux'], $log->getMetaData());
 
-                return true;
-            }));
+                    return true;
+                })
+            );
 
         /**
          * @var WriterInterface   $writer
          * @var PriorityInterface $priority
+         * @var Throwable         $throwable
          */
         $logger = new Logger();
         $logger
             ->addWriter($writer)
             ->addDecorator($decorator)
-            ->log('Error!', $priority, ['foo' => 'bar']);
+            ->log('Error!', $priority, $throwable, ['foo' => 'bar']);
     }
 
     /**
@@ -72,6 +79,7 @@ class LoggerTest extends TestCase
     public function testInterrupt(): void
     {
         $priority = $this->createMock(PriorityInterface::class);
+        $throwable = $this->createMock(Throwable::class);
 
         $writer = $this->createMock(WriterInterface::class);
         $writer
@@ -81,12 +89,13 @@ class LoggerTest extends TestCase
         /**
          * @var WriterInterface   $writer
          * @var PriorityInterface $priority
+         * @var Throwable         $throwable
          */
         $logger = new Logger();
         $logger
             ->addWriter($writer, true)
             ->addWriter($writer)
             ->addWriter($writer)
-            ->log('Error!', $priority, ['foo' => 'bar']);
+            ->log('Error!', $priority, $throwable, ['foo' => 'bar']);
     }
 }